+++ Mods ChoGGi/Autonomous Extractors/Code/Script.lua
@@ -1,14 +1,16 @@
 -- See LICENSE for terms
 
 local ToggleWorking = ChoGGi.ComFuncs.ToggleWorking
+
 local mod_AutoPerformance
+local mod_EnableMod
 
 local has_colonists = {
 	MetalsExtractor = true,
 	PreciousMetalsExtractor = true,
 }
 
-local function StartupCode()
+local function UpdateBuildings()
 
 	-- make newly built buildings not need workers
 	local bt = BuildingTemplates
@@ -16,33 +18,47 @@ local function StartupCode()
 
 	for id, item in pairs(bt) do
 		if has_colonists[id] then
-			item.max_workers = 0
-			item.automation = 1
-			item.auto_performance = mod_AutoPerformance
-			local cls_temp = ct[id]
-			cls_temp.max_workers = 0
-			cls_temp.automation = 1
-			cls_temp.auto_performance = mod_AutoPerformance
+			if mod_EnableMod then
+				item.max_workers = 0
+				item.automation = 1
+				item.auto_performance = mod_AutoPerformance
+				local cls_temp = ct[id]
+				cls_temp.max_workers = 0
+				cls_temp.automation = 1
+				cls_temp.auto_performance = mod_AutoPerformance
+			else
+				item.max_workers = nil
+				item.automation = nil
+				item.auto_performance = nil
+				local cls_temp = ct[id]
+				cls_temp.max_workers = nil
+				cls_temp.automation = nil
+				cls_temp.auto_performance = nil
+			end
 		end
 	end
 
 	-- update existing buildings
 	local objs = UIColony:GetCityLabels("MetalsExtractor")
-	-- no need to update again
-	if objs[1] and objs[1].max_workers ~= 0 then
-		table.iappend(objs, UIColony:GetCityLabels("PreciousMetalsExtractor"))
-		for i = 1, #objs do
-			local obj = objs[i]
+	table.iappend(objs, UIColony:GetCityLabels("PreciousMetalsExtractor"))
+	for i = 1, #objs do
+		local obj = objs[i]
+
+		if mod_EnableMod then
 			obj.max_workers = 0
 			obj.automation = 1
 			obj.auto_performance = mod_AutoPerformance
-			ToggleWorking(obj)
+		else
+			obj.max_workers = nil
+			obj.automation = nil
+			obj.auto_performance = nil
 		end
-	end
 
+		ToggleWorking(obj)
+	end
 end
-OnMsg.CityStart = StartupCode
-OnMsg.LoadGame = StartupCode
+OnMsg.CityStart = UpdateBuildings
+OnMsg.LoadGame = UpdateBuildings
 
 local function ModOptions(id)
 	-- id is from ApplyModOptions
@@ -50,6 +66,7 @@ local function ModOptions(id)
 		return
 	end
 
+	mod_EnableMod = CurrentModOptions:GetProperty("EnableMod")
 	mod_AutoPerformance = CurrentModOptions:GetProperty("AutoPerformance")
 
 	-- Make sure we're in-game
@@ -57,15 +74,7 @@ local function ModOptions(id)
 		return
 	end
 
-	StartupCode()
-
-	-- update existing buildings auto_performance
-	local objs = UIColony:GetCityLabels("MetalsExtractor")
-	table.iappend(objs, UIColony:GetCityLabels("PreciousMetalsExtractor"))
-	for i = 1, #objs do
-		objs[i].auto_performance = mod_AutoPerformance
-	end
-
+	UpdateBuildings()
 end
 -- Load default/saved settings
 OnMsg.ModsReloaded = ModOptions