+++ Mods ChoGGi/Game Rule Amazonian Mars/items.lua
@@ -1,0 +1,12 @@
+-- See LICENSE for terms
+
+return {
+	PlaceObj("ModItemOptionNumber", {
+		"name", "AmazonianSize",
+		"DisplayName", T(0000, "Amazonian Size"),
+		"Help", T(0000, "Extra snu-snu."),
+		"DefaultValue", 100,
+		"MinValue", 1,
+		"MaxValue", 1000,
+	}),
+}