+++ Mods ChoGGi/Game Rule Amazonian Mars/metadata.lua
@@ -1,0 +1,25 @@
+return PlaceObj("ModDef", {
+	"title", "Game Rule Amazonian Mars",
+	"id", "ChoGGi_GameRuleAmazonianMars",
+	"steam_id", "2856628836",
+	"pops_any_uuid", "3042c4ec-167e-499d-82f0-76d6f55b5486",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+You may only import female colonists (exception: tourists).
+Male tourists "provide" children (Snu-snu, mod option for larger female colonist models).
+Non-female children are culled at birth (lower birth rate).
+and just in case a daily cull of any non-tourist non-female colonists on Mars (mini-soylent green).
+
+If the [url=https://steamcommunity.com/sharedfiles/filedetails/?id=1442280139]Planet Express Colony Logo[/url] mod is installed than any rockets with a tourist use that logo.
+]],
+})