+++ Mods ChoGGi/Lock Workplace/items.lua
@@ -7,4 +7,10 @@ return {
 		"Help", T(302535920011579, "Workers will never change workplace (may cause issues)."),
 		"DefaultValue", false,
 	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "SeniorsOverride",
+		"DisplayName", T(0000, "Seniors Override"),
+		"Help", T(0000, "Allow graceful retirement."),
+		"DefaultValue", true,
+	}),
 }