+++ Mods ChoGGi/Lower Maintenance/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2842441599",
 	"pops_any_uuid", "97cf86d1-7add-440a-bb64-562277ee7652",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {